apps/imageAI/requirements.txt
flask
tensorflow==1.14
keras==2.3.1
imageai
h5py
pillow